requirements.txt
langchain
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit
openai==0.27.6
#faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
streamlit_extras
spacy
google-generativeai
langchain_google_genai
nltk
wordcloud
faiss-cpu
grpcio
langchain-community

# below also is required.uncomment as it is required.

#python -m spacy download en_core_web_sm